Revert "refactor: page server component"
This reverts commit 298b2869c0210d41f87bfa7738c79045981ed9ce.
