+++ packages/mesh/app/social/_components/no-followings.tsx
@@ -1,5 +1,3 @@
-'use client'
-
 import { useEffect, useState } from 'react'
 
 import { getMostFollowersData } from '@/app/actions/get-member-followings'
