+++ packages/mesh/app/social/page.tsx
@@ -1,25 +1,47 @@
-import { notFound, redirect } from 'next/navigation'
+'use client'
+
+import { useEffect, useState } from 'react'
 
 import { getSocialPageData } from '@/app/actions/get-member-followings'
+import ErrorPage from '@/components/status/error-page'
+import { useUser } from '@/context/user'
+import { type MongoDBResponse } from '@/utils/data-schema'
 
-import { getCurrentUser } from '../actions/auth'
 import Feed from './_components/feed'
 import FollowSuggestionFeed from './_components/follow-suggestion-feed'
 import FollowSuggestionWidget from './_components/follow-suggestion-widget'
+import Loading from './_components/loading'
 import NoFollowings from './_components/no-followings'
 
-export default async function Page() {
+export default function Page() {
   //TODO: infiniteScroll
+  const { user } = useUser()
   // const feedsNumber = 20
   const firstSectionAmount = 3
   const suggestedFollowersNumber = 5
-  const user = await getCurrentUser()
-
-  if (!user) redirect('/login')
   const memberId = user.memberId
+  const [socialData, setSocialData] = useState<MongoDBResponse | null>(null)
+  const [isLoading, setIsLoading] = useState(false)
+  const [isNotFound, setIsNotFound] = useState(false)
 
-  const socialData = await getSocialPageData(memberId)
-  if (!socialData) notFound()
+  useEffect(() => {
+    const fetchSocialData = async () => {
+      setIsLoading(true)
+      const response = await getSocialPageData(memberId)
+      if (!response) {
+        setIsNotFound(true)
+      } else {
+        setSocialData(response)
+      }
+      setIsLoading(false)
+    }
+    fetchSocialData()
+  }, [memberId])
+
+  if (isLoading) return <Loading />
+  if (isNotFound) return <ErrorPage statusCode={404} />
+  if (!socialData) return <Loading />
+
   const { stories, members } = socialData
 
   if (!members.length && !stories.length) {
